fix(codex): answer startup probes before first attach
